services/python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
prometheus-client==0.20.0
kss==6.0.5
nats-py>=2.8.0
pydantic==2.8.2
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
httpx
deepl
asyncpg